Predeclare spotless deps instead
See https://github.com/diffplug/spotless/issues/1213. Better performance too since we use the same config for every project
